Renamed IGreetingService -> GreetingService, GreetingService -> GreetingServiceImpl
