+++ src/main/java/org/noteskeeper/server/GreetingService.java
@@ -1,16 +1,5 @@
 package org.noteskeeper.server;
 
-import org.springframework.stereotype.Service;
-
-@Service("greetingService")
-public class GreetingService implements IGreetingService{
-
-    @Override
-    public String greet(String name){
-        if (name == null || name.isEmpty()) {
-            return "Hello, Anonymous";
-        } else {
-            return "Hello, " + name;
-        }
-    }
+public interface GreetingService {
+     String greet(String name);
 }
