--- src/main/java/org/noteskeeper/server/IGreetingService.java
@@ -1,5 +1,0 @@
-package org.noteskeeper.server;
-
-public interface IGreetingService {
-     String greet(String name);
-}
