+++ src/main/webapp/WEB-INF/applicationContext.xml
@@ -9,6 +9,6 @@
            http://www.springframework.org/schema/context/spring-context-2.5.xsd">
 
     <context:component-scan base-package="server" />
-    <bean id="greetingService" class="org.noteskeeper.server.GreetingService" />
+    <bean id="greetingService" class="org.noteskeeper.server.GreetingServiceImpl" />
     <bean class="org.noteskeeper.server.HelloWorld"/>
 </beans>
